adjust unittest
